Repackage dependencies
