+++ buildSrc/src/main/kotlin/io/spine/internal/dependency/AppEngine.kt
@@ -24,21 +24,13 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import io.spine.internal.dependency.JUnit
-import io.spine.internal.dependency.Truth
+package io.spine.internal.dependency
 
-group = "io.spine.tools"
+// https://cloud.google.com/java/docs/reference
+object AppEngine {
+    private const val version = "1.9.82"
+    private const val gradlePluginVersion = "2.2.0"
 
-val spineBaseVersion: String by extra
-
-dependencies {
-    api(project(":client"))
-    api("io.spine.tools:spine-testlib:$spineBaseVersion")
-
-    JUnit.api.forEach {
-        api(it)
-    }
-    Truth.libs.forEach {
-       api(it)
-    }
+    const val sdk          = "com.google.appengine:appengine-api-1.0-sdk:${version}"
+    const val gradlePlugin = "com.google.cloud.tools:appengine-gradle-plugin:${gradlePluginVersion}"
 }
