+++ buildSrc/src/main/kotlin/io/spine/internal/dependency/CheckStyle.kt
@@ -24,21 +24,11 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import io.spine.internal.dependency.JUnit
-import io.spine.internal.dependency.Truth
+package io.spine.internal.dependency
 
-group = "io.spine.tools"
-
-val spineBaseVersion: String by extra
-
-dependencies {
-    api(project(":client"))
-    api("io.spine.tools:spine-testlib:$spineBaseVersion")
-
-    JUnit.api.forEach {
-        api(it)
-    }
-    Truth.libs.forEach {
-       api(it)
-    }
+// https://checkstyle.sourceforge.io/
+// See `config/gradle.checkstyle.gradle`.
+@Suppress("unused")
+object CheckStyle {
+    const val version = "8.29"
 }
