+++ buildSrc/src/main/kotlin/io/spine/internal/dependency/Firebase.kt
@@ -24,21 +24,10 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import io.spine.internal.dependency.JUnit
-import io.spine.internal.dependency.Truth
+package io.spine.internal.dependency
 
-group = "io.spine.tools"
-
-val spineBaseVersion: String by extra
-
-dependencies {
-    api(project(":client"))
-    api("io.spine.tools:spine-testlib:$spineBaseVersion")
-
-    JUnit.api.forEach {
-        api(it)
-    }
-    Truth.libs.forEach {
-       api(it)
-    }
+// https://firebase.google.com/docs/admin/setup#java
+object Firebase {
+    private const val adminVersion = "6.12.2"
+    const val admin = "com.google.firebase:firebase-admin:${adminVersion}"
 }
