+++ buildSrc/src/main/kotlin/io/spine/internal/dependency/Flogger.kt
@@ -24,21 +24,16 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import io.spine.internal.dependency.JUnit
-import io.spine.internal.dependency.Truth
+package io.spine.internal.dependency
 
-group = "io.spine.tools"
-
-val spineBaseVersion: String by extra
-
-dependencies {
-    api(project(":client"))
-    api("io.spine.tools:spine-testlib:$spineBaseVersion")
-
-    JUnit.api.forEach {
-        api(it)
-    }
-    Truth.libs.forEach {
-       api(it)
+// https://github.com/google/flogger
+object Flogger {
+    internal const val version = "0.6"
+    const val lib     = "com.google.flogger:flogger:${version}"
+    @Suppress("unused")
+    object Runtime {
+        const val systemBackend = "com.google.flogger:flogger-system-backend:${version}"
+        const val log4J         = "com.google.flogger:flogger-log4j:${version}"
+        const val slf4J         = "com.google.flogger:slf4j-backend-factory:${version}"
     }
 }
