+++ buildSrc/src/main/kotlin/io/spine/internal/gradle/GradlePlugins.kt
@@ -24,21 +24,15 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import io.spine.internal.dependency.JUnit
-import io.spine.internal.dependency.Truth
+package io.spine.internal.gradle
 
-group = "io.spine.tools"
+import io.spine.internal.dependency.AppEngine
+import io.spine.internal.dependency.ErrorProne
+import io.spine.internal.dependency.Protobuf
 
-val spineBaseVersion: String by extra
-
-dependencies {
-    api(project(":client"))
-    api("io.spine.tools:spine-testlib:$spineBaseVersion")
-
-    JUnit.api.forEach {
-        api(it)
-    }
-    Truth.libs.forEach {
-       api(it)
-    }
+@Suppress("unused")
+object GradlePlugins {
+    const val errorProne  = ErrorProne.gradlePlugin
+    const val protobuf    = Protobuf.gradlePlugin
+    const val appengine   = AppEngine.gradlePlugin
 }
