+++ buildSrc/src/main/kotlin/io/spine/internal/gradle/Repositories.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.gradle.internal
+package io.spine.internal.gradle
 
 import java.io.File
 import java.net.URI
